requirements.txt
Flask~=1.1.2
pymongo~=3.11.3
python-dotenv~=0.17.0